refactor: change to CRA setup
