--- .babelrc
@@ -1,4 +1,0 @@
-{
-  "presets": ["env", "react"],
-  "plugins": ["transform-class-properties", "transform-object-rest-spread"]
-}
