--- babelcmd.txt
@@ -1,1 +1,0 @@
-babel src/app.js --out-file=public/scripts/app.js --presets=env,react --watch
