--- jest.config.json
@@ -1,4 +1,0 @@
-{
-  "setupFiles": ["raf/polyfill", "<rootDir>/src/tests/setupTests.js"],
-  "snapshotSerializers": ["enzyme-to-json/serializer"]
-}
