+++ package.json
@@ -1,33 +1,17 @@
 {
-  "name": "meditation-journal-app",
-  "version": "1.0.0",
-  "main": "index.js",
-  "author": "Sam MacFarlane",
-  "license": "MIT",
-  "scripts": {
-    "build:dev": "webpack",
-    "build:prod": "webpack -p --env production",
-    "dev-server": "webpack-dev-server",
-    "test": "cross-env NODE_ENV=test jest --config=jest.config.json",
-    "test:watch": "yarn test --watch",
-    "heroku-postbuild": "yarn run build:prod"
-  },
+  "name": "react-meditation-journal-v2",
+  "version": "0.1.0",
+  "private": true,
   "dependencies": {
     "@fortawesome/fontawesome-svg-core": "^1.2.28",
     "@fortawesome/free-solid-svg-icons": "^5.13.0",
     "@fortawesome/react-fontawesome": "^0.1.9",
     "@material-ui/core": "^4.10.1",
     "@react-mock/state": "^0.1.8",
-    "@testing-library/react": "^11.0.4",
-    "@testing-library/user-event": "^12.1.7",
-    "babel-cli": "6.24.1",
-    "babel-core": "6.25.0",
-    "babel-loader": "7.1.1",
-    "babel-plugin-transform-class-properties": "6.24.1",
-    "babel-plugin-transform-object-rest-spread": "6.23.0",
-    "babel-polyfill": "6.26.0",
-    "babel-preset-env": "1.5.2",
-    "babel-preset-react": "6.24.1",
+    "@testing-library/jest-dom": "^5.11.4",
+    "@testing-library/react": "^11.1.0",
+    "@testing-library/user-event": "^12.1.10",
+    "@wojtekmaj/enzyme-adapter-react-17": "^0.6.1",
     "css-loader": "0.28.4",
     "enzyme": "^3.11.0",
     "express": "^4.17.1",
@@ -39,13 +23,14 @@
     "node-sass": "4.13.1",
     "normalize.css": "7.0.0",
     "raf": "3.3.2",
-    "react": "^16.13.1",
+    "react": "16.8.6",
     "react-addons-shallow-compare": "15.6.0",
     "react-dates": "12.7.0",
-    "react-dom": "^16.13.1",
+    "react-dom": "16.8.6",
     "react-modal": "^3.11.2",
     "react-redux": "^7.2.1",
     "react-router-dom": "^5.2.0",
+    "react-scripts": "4.0.3",
     "redux": "^4.0.5",
     "redux-mock-store": "^1.5.4",
     "redux-thunk": "2.2.0",
@@ -53,7 +38,7 @@
     "style-loader": "0.18.2",
     "uuid": "3.1.0",
     "validator": "8.0.0",
-    "webpack": "3.1.0",
+    "web-vitals": "^1.0.1",
     "yarn": "^1.22.4"
   },
   "devDependencies": {
@@ -61,10 +46,30 @@
     "cross-env": "5.0.5",
     "dotenv": "4.0.0",
     "enzyme-adapter-react-16": "^1.15.5",
-    "enzyme-to-json": "3.0.1",
-    "eslint": "^7.10.0",
-    "jest": "20.0.4",
-    "react-test-renderer": "16.0.0",
-    "webpack-dev-server": "2.5.1"
+    "enzyme-to-json": "3.0.1"
+  },
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
