+++ src/tests/__mocks__/moment.js
@@ -1,4 +1,4 @@
-const moment = require.requireActual('moment');
+const moment = require.requireActual("moment");
 
 export default (timestamp = 0) => {
   return moment(timestamp);
