+++ src/tests/setupTests.js
@@ -1,9 +1,9 @@
-import Enzyme from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
-import DotEnv from 'dotenv';
+import Enzyme from "enzyme";
+import Adapter from "enzyme-adapter-react-16";
+import DotEnv from "dotenv";
 
 Enzyme.configure({
   adapter: new Adapter(),
 });
 
-DotEnv.config({ path: '.env.test' });
+DotEnv.config({ path: ".env.test" });
